+++ guillotina/files/adapter.py
@@ -1,5 +1,4 @@
 from .dbfile import DBFile
-from aiohttp.web_exceptions import HTTPPreconditionFailed
 from guillotina import configure
 from guillotina.blob import Blob
 from guillotina.event import notify
@@ -13,6 +12,7 @@ from guillotina.interfaces import IFileStorageManager
 from guillotina.interfaces import IRequest
 from guillotina.interfaces import IResource
 from guillotina.interfaces import IUploadDataManager
+from guillotina.response import HTTPPreconditionFailed
 
 import time